Move VALUE_CLICK_TRIGGER and APPLY_FILTER_TRIGGER to ui_action… (#60202)
* move triggers

* move triggers

* fix import path

* fix import path
