+++ src/legacy/core_plugins/kibana/public/discover/np_ready/embeddable/search_embeddable.ts
@@ -20,7 +20,10 @@ import _ from 'lodash';
 import * as Rx from 'rxjs';
 import { Subscription } from 'rxjs';
 import { i18n } from '@kbn/i18n';
-import { UiActionsStart } from 'src/plugins/ui_actions/public';
+import {
+  UiActionsStart,
+  APPLY_FILTER_TRIGGER,
+} from '../../../../../../..//plugins/ui_actions/public';
 import { RequestAdapter, Adapters } from '../../../../../../../plugins/inspector/public';
 import {
   esFilters,
@@ -31,11 +34,7 @@ import {
   Query,
   IFieldType,
 } from '../../../../../../../plugins/data/public';
-import {
-  APPLY_FILTER_TRIGGER,
-  Container,
-  Embeddable,
-} from '../../../../../embeddable_api/public/np_ready/public';
+import { Container, Embeddable } from '../../../../../embeddable_api/public/np_ready/public';
 import * as columnActions from '../angular/doc_table/actions/columns';
 import searchTemplate from './search_template.html';
 import { ISearchEmbeddable, SearchInput, SearchOutput } from './types';
