+++ src/plugins/embeddable/public/bootstrap.ts
@@ -17,18 +17,11 @@
  * under the License.
  */
 import { UiActionsSetup } from '../../ui_actions/public';
-import { Filter } from '../../data/public';
 import {
-  applyFilterTrigger,
   contextMenuTrigger,
   createFilterAction,
   panelBadgeTrigger,
-  valueClickTrigger,
-  EmbeddableVisTriggerContext,
-  IEmbeddable,
   EmbeddableContext,
-  APPLY_FILTER_TRIGGER,
-  VALUE_CLICK_TRIGGER,
   CONTEXT_MENU_TRIGGER,
   PANEL_BADGE_TRIGGER,
   ACTION_ADD_PANEL,
@@ -42,11 +35,6 @@ import {
 
 declare module '../../ui_actions/public' {
   export interface TriggerContextMapping {
-    [VALUE_CLICK_TRIGGER]: EmbeddableVisTriggerContext;
-    [APPLY_FILTER_TRIGGER]: {
-      embeddable: IEmbeddable;
-      filters: Filter[];
-    };
     [CONTEXT_MENU_TRIGGER]: EmbeddableContext;
     [PANEL_BADGE_TRIGGER]: EmbeddableContext;
   }
@@ -67,9 +55,7 @@ declare module '../../ui_actions/public' {
  */
 export const bootstrap = (uiActions: UiActionsSetup) => {
   uiActions.registerTrigger(contextMenuTrigger);
-  uiActions.registerTrigger(applyFilterTrigger);
   uiActions.registerTrigger(panelBadgeTrigger);
-  uiActions.registerTrigger(valueClickTrigger);
 
   const actionApplyFilter = createFilterAction();
