+++ src/plugins/embeddable/public/lib/triggers/triggers.ts
@@ -33,25 +33,11 @@ export interface EmbeddableVisTriggerContext {
   };
 }
 
-export const VALUE_CLICK_TRIGGER = 'VALUE_CLICK_TRIGGER';
-export const valueClickTrigger: Trigger<'VALUE_CLICK_TRIGGER'> = {
-  id: VALUE_CLICK_TRIGGER,
-  title: 'Value clicked',
-  description: 'Value was clicked',
-};
-
 export const CONTEXT_MENU_TRIGGER = 'CONTEXT_MENU_TRIGGER';
 export const contextMenuTrigger: Trigger<'CONTEXT_MENU_TRIGGER'> = {
   id: CONTEXT_MENU_TRIGGER,
   title: 'Context menu',
   description: 'Triggered on top-right corner context-menu select.',
-};
-
-export const APPLY_FILTER_TRIGGER = 'FILTER_TRIGGER';
-export const applyFilterTrigger: Trigger<'FILTER_TRIGGER'> = {
-  id: APPLY_FILTER_TRIGGER,
-  title: 'Filter click',
-  description: 'Triggered when user applies filter to an embeddable.',
 };
 
 export const PANEL_BADGE_TRIGGER = 'PANEL_BADGE_TRIGGER';
